c++20: Fix division of a double by an enum.
The c++20 standard has deprecated automatic conversions from an enum
to a float.  The CD_FRAMESAMPLES value probably shouldn't be part of
an enum in the first place, but that's the way it is provided by the
cdda headers.  Add an explicit cast to this value to an int, and let
the compiler perform its usual int->float conversion.

+++ mythplugins/mythmusic/mythmusic/cddecoder.cpp
@@ -322,7 +322,7 @@ void CdDecoder::run()
         if (m_seekTime >= +0.)
         {
             m_curPos = m_start + static_cast< lsn_t >(
-                (m_seekTime * kSamplesPerSec) / CD_FRAMESAMPLES);
+                (m_seekTime * kSamplesPerSec) / static_cast<int>(CD_FRAMESAMPLES));
             if (m_paranoia)
             {
                 QMutexLocker lock(&getCdioMutex());
